# Conflicts:
#	src/main/java/io/github/mzmine/modules/batchmode/BatchComponentController.java
#	src/main/java/io/github/mzmine/modules/batchmode/BatchModeModulesList.java
#	src/main/java/io/github/mzmine/modules/dataprocessing/gapfill_peakfinder/PeakFinderModule.java
#	src/main/java/io/github/mzmine/modules/tools/sortfeaturelists/SortFeatureListsModule.java
#	src/main/java/io/github/mzmine/parameters/dialogs/ParameterSetupDialog.java

+++ src/main/java/io/github/mzmine/modules/io/projectsave/ProjectSavingTask.java
@@ -1,19 +1,26 @@
 /*
- * Copyright 2006-2021 The MZmine Development Team
+ * Copyright (c) 2004-2023 The MZmine Development Team
  *
- * This file is part of MZmine.
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
  *
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
  *
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
  */
 
 package io.github.mzmine.modules.io.projectsave;
@@ -23,24 +30,26 @@ import io.github.mzmine.datamodel.RawDataFile;
 import io.github.mzmine.datamodel.features.FeatureList;
 import io.github.mzmine.datamodel.features.ModularFeatureList;
 import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.modules.io.projectload.ProjectLoaderParameters;
 import io.github.mzmine.parameters.ParameterSet;
 import io.github.mzmine.project.impl.MZmineProjectImpl;
 import io.github.mzmine.taskcontrol.AbstractTask;
 import io.github.mzmine.taskcontrol.TaskStatus;
 import io.github.mzmine.util.ExceptionUtils;
 import io.github.mzmine.util.StreamCopy;
+import io.github.mzmine.util.files.FileAndPathUtil;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Date;
+import java.time.Instant;
 import java.util.Hashtable;
+import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
-import javafx.scene.control.ButtonType;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerConfigurationException;
 import org.jetbrains.annotations.NotNull;
@@ -49,38 +58,34 @@ import org.xml.sax.SAXException;
 public class ProjectSavingTask extends AbstractTask {
 
   public static final String VERSION_FILENAME = "MZMINE_VERSION";
+  public static final String STANDALONE_FILENAME = "STANDALONE"; // only exists if it's a standalone project.
   public static final String CONFIG_FILENAME = "configuration.xml";
   public static final String PARAMETERS_FILENAME = "User parameters.xml";
+  private static final Logger logger = Logger.getLogger(ProjectSavingTask.class.getName());
+  private final ProjectSaveOption projectType;
 
-  private Logger logger = Logger.getLogger(this.getClass().getName());
-
-  private File saveFile;
-  private MZmineProjectImpl savedProject;
-
+  private final File saveFile;
+  private final MZmineProjectImpl savedProject;
+  private final int totalSaveItems;
+  private final int finishedSaveItems = 0;
+  // This hashtable maps raw data files to their ID within the saved project
+  private final Hashtable<RawDataFile, String> dataFilesIDMap;
   private RawDataFileSaveHandler rawDataFileSaveHandler;
   private PeakListSaveHandler peakListSaveHandler;
   private UserParameterSaveHandler userParameterSaveHandler;
-
-  private final int totalSaveItems;
-  private int currentStage, finishedSaveItems = 0;
+  private int currentStage;
   private String currentSavedObjectName;
 
-  // This hashtable maps raw data files to their ID within the saved project
-  private Hashtable<RawDataFile, String> dataFilesIDMap;
-  private boolean saveStandaloneProject = true;
-
   public ProjectSavingTask(MZmineProject project, ParameterSet parameters,
-      @NotNull Date moduleCallDate) {
+      @NotNull Instant moduleCallDate) {
     super(null, moduleCallDate);
     this.savedProject = (MZmineProjectImpl) project;
-    this.saveFile = parameters.getParameter(ProjectLoaderParameters.projectFile).getValue();
-    dataFilesIDMap = new Hashtable<RawDataFile, String>();
-    this.totalSaveItems = project.getDataFiles().length + project.getFeatureLists().size();
+    this.saveFile = parameters.getValue(ProjectSaveAsParameters.projectFile);
+    this.projectType = parameters.getValue(ProjectSaveAsParameters.option);
+    dataFilesIDMap = new Hashtable<>();
+    this.totalSaveItems = project.getDataFiles().length + project.getCurrentFeatureLists().size();
   }
 
-  /**
-   * @see io.github.mzmine.taskcontrol.Task#getTaskDescription()
-   */
   @Override
   public String getTaskDescription() {
     if (currentSavedObjectName == null) {
@@ -89,9 +94,6 @@ public class ProjectSavingTask extends AbstractTask {
     return "Saving project (" + currentSavedObjectName + ")";
   }
 
-  /**
-   * @see io.github.mzmine.taskcontrol.Task#getFinishedPercentage()
-   */
   @Override
   public double getFinishedPercentage() {
 
@@ -119,14 +121,9 @@ public class ProjectSavingTask extends AbstractTask {
         return 0;
     }
 
-    double progress = (finishedSaveItems + currentItemProgress) / totalSaveItems;
-
-    return progress;
+    return (finishedSaveItems + currentItemProgress) / totalSaveItems;
   }
 
-  /**
-   * @see io.github.mzmine.taskcontrol.Task#cancel()
-   */
   @Override
   public void cancel() {
 
@@ -148,37 +145,28 @@ public class ProjectSavingTask extends AbstractTask {
 
   }
 
-  /**
-   * @see java.lang.Runnable#run()
-   */
   @Override
   public void run() {
     try {
       logger.info("Saving project to " + saveFile);
       setStatus(TaskStatus.PROCESSING);
 
-      if (!MZmineCore.isHeadLessMode()) {
-        ButtonType btn = MZmineCore.getDesktop().displayConfirmation(
-            "Would you like to save a standalone project?\n\nIf yes, raw data files will be "
-                + "copied into the project file. Otherwise the project will require that all raw "
-                + "files stay in the same directories.\n<b>WARNING</b> If this is an existing "
-                + "project, it is recommended to save it in the same way.", ButtonType.YES,
-            ButtonType.NO, ButtonType.CANCEL);
-        if (btn == ButtonType.CANCEL) {
-          setStatus(TaskStatus.FINISHED);
-          return;
-        } else if (btn == ButtonType.YES) {
-          saveStandaloneProject = true;
-        } else {
-          saveStandaloneProject = false;
-        }
-      } else {
-        saveStandaloneProject = true;
+      switch (projectType) {
+        case STANDALONE -> savedProject.setStandalone(true);
+        case REFERENCING -> savedProject.setStandalone(false);
       }
 
       // Prepare a temporary ZIP file. We create this file in the same
       // directory as the final saveFile to avoid moving between
       // filesystems in the last stage (renameTo)
+
+      // checking if directory exists first
+      if (!FileAndPathUtil.createDirectory(saveFile.getParentFile())) {
+        setErrorMessage("Could not create directories for file " + saveFile + " for writing.");
+        setStatus(TaskStatus.ERROR);
+        return;
+      }
+
       File tempFile = File.createTempFile(saveFile.getName(), ".tmp", saveFile.getParentFile());
       tempFile.deleteOnExit();
 
@@ -189,6 +177,7 @@ public class ProjectSavingTask extends AbstractTask {
       // Stage 1 - save version and configuration
       currentStage++;
       saveVersion(zipStream);
+      saveStandalone(zipStream);
       saveConfiguration(zipStream);
       if (isCanceled()) {
         zipStream.close();
@@ -262,8 +251,7 @@ public class ProjectSavingTask extends AbstractTask {
 
     } catch (Throwable e) {
 
-      e.printStackTrace();
-
+      logger.log(Level.SEVERE, e.getMessage(), e);
       setStatus(TaskStatus.ERROR);
 
       if (currentSavedObjectName == null) {
@@ -292,6 +280,14 @@ public class ProjectSavingTask extends AbstractTask {
 
   }
 
+  private void saveStandalone(ZipOutputStream zipStream) throws IOException {
+    if (savedProject.isStandalone()) {
+      zipStream.putNextEntry(new ZipEntry(STANDALONE_FILENAME));
+      String MZmineVersion = STANDALONE_FILENAME;
+      zipStream.write(MZmineVersion.getBytes());
+    }
+  }
+
   /**
    * Save the configuration file.
    *
@@ -306,7 +302,7 @@ public class ProjectSavingTask extends AbstractTask {
     zipStream.putNextEntry(new ZipEntry(CONFIG_FILENAME));
 
     try {
-      File tempConfigFile = File.createTempFile("mzmineconfig", ".tmp");
+      File tempConfigFile = FileAndPathUtil.createTempFile("mzmineconfig", ".tmp");
       MZmineCore.getConfiguration().saveConfiguration(tempConfigFile);
       FileInputStream fileStream = new FileInputStream(tempConfigFile);
 
@@ -316,7 +312,7 @@ public class ProjectSavingTask extends AbstractTask {
       fileStream.close();
       tempConfigFile.delete();
     } catch (Exception e) {
-      e.printStackTrace();
+      logger.log(Level.SEVERE, e.getMessage(), e);
       logger.warning("Could not save configuration" + ExceptionUtils.exceptionToString(e));
     }
   }
@@ -329,7 +325,7 @@ public class ProjectSavingTask extends AbstractTask {
 
     AtomicBoolean finished = new AtomicBoolean(false);
     rawDataFileSaveHandler = new RawDataFileSaveHandler(savedProject, zipStream,
-        saveStandaloneProject, getModuleCallDate());
+        Objects.requireNonNullElse(savedProject.isStandalone(), true), getModuleCallDate());
     rawDataFileSaveHandler.addTaskStatusListener((task, newStatus, oldStatus) -> {
       switch (newStatus) {
         case WAITING, PROCESSING -> {
@@ -354,7 +350,7 @@ public class ProjectSavingTask extends AbstractTask {
       try {
         Thread.sleep(100);
       } catch (InterruptedException e) {
-        e.printStackTrace();
+        logger.log(Level.SEVERE, e.getMessage(), e);
       }
     }
   }
@@ -368,7 +364,8 @@ public class ProjectSavingTask extends AbstractTask {
   private void savePeakLists(ZipOutputStream zipStream)
       throws IOException, TransformerConfigurationException, SAXException {
 
-    for (FeatureList featureList : savedProject.getFeatureLists()) {
+    final List<FeatureList> currentFeatureLists = savedProject.getCurrentFeatureLists();
+    for (FeatureList featureList : currentFeatureLists) {
       FeatureListSaveTask saveTask = new FeatureListSaveTask((ModularFeatureList) featureList,
           zipStream);
 
@@ -388,7 +385,7 @@ public class ProjectSavingTask extends AbstractTask {
         try {
           Thread.sleep(50);
         } catch (InterruptedException e) {
-          e.printStackTrace();
+          logger.log(Level.SEVERE, e.getMessage(), e);
         }
       }
